+++ R/print.netmeta.R
@@ -371,23 +371,22 @@ print.netmeta <- function(x,
     if (reference.group != "") {
       if (baseline.reference)
         comptext <-
-          paste0("comparison: ",
-                 if (x$n == 2)
-                   paste0("'",
-                          treats(rownames(TE.common),
-                                 nchar.trts)[rownames(TE.common)
-                                             != reference.group],
-                          "'")
-                 else
-                   "other treatments",
-                 " vs '",
-                 treats(rownames(TE.common),
-                        nchar.trts)[rownames(TE.common)
-                                    == reference.group],
-                 "'")
+          paste0(if (x$n == 2)
+            paste0("'",
+                   treats(rownames(TE.common),
+                          nchar.trts)[rownames(TE.common)
+                                      != reference.group],
+                   "'")
+            else
+              "other treatments",
+            " vs '",
+            treats(rownames(TE.common),
+                   nchar.trts)[rownames(TE.common)
+                               == reference.group],
+            "'")
       else
         comptext <-
-          paste0("comparison: '",
+          paste0("'",
                  treats(rownames(TE.common),
                         nchar.trts)[rownames(TE.common)
                                     == reference.group],
@@ -568,8 +567,7 @@ print.netmeta <- function(x,
         ##
         rownames(res) <- treats(rownames(res), nchar.trts)
         ##
-        cat("\nTreatment estimate (sm = '", sm.lab,
-            "', ", comptext, "):\n", sep = "")
+        cat("\nTreatment estimate (", comptext, "):\n", sep = "")
         ##
         prmatrix(res, quote = FALSE, right = TRUE)
       }
@@ -734,8 +732,7 @@ print.netmeta <- function(x,
         ##
         rownames(res) <- treats(rownames(res), nchar.trts)
         ##
-        cat("\nTreatment estimate (sm = '", sm.lab,
-            "', ", comptext, "):\n", sep = "")
+        cat("\nTreatment estimate (", comptext, "):\n", sep = "")
         ##
         prmatrix(res, quote = FALSE, right = TRUE)
       }