+++ R/netrank.R
@@ -13,7 +13,7 @@
 #' @param method A character string specifying whether the
 #'   \code{"P-score"} or \code{"SUCRA"} ranking metric will be
 #'   calculated.
-#' @param nsim Number of simulations to calculate SUCRAs.
+#' @param nsim Number of repetitions to calculate SUCRAs.
 #' @param common A logical indicating whether to print P-scores or
 #'   SUCRAs for the common effects model.
 #' @param random A logical indicating whether to print P-scores or
@@ -41,7 +41,7 @@
 #' of treatment \emph{i} within the range of treatments, measured on a
 #' scale from 0 (worst) to 1 (best) (Salanti et al. 2011). A
 #' resampling method is used to calculate SUCRAs for frequentist
-#' network meta-analysis. The number of simulations is determine by
+#' network meta-analysis. The number of repetitions is determine by
 #' argument \code{nsim}.
 #'
 #' The interpretation of P-scores and SUCRAs is comparable.
@@ -523,7 +523,7 @@ print.netrank <- function(x,
 
   if (x$method == "SUCRA" & !is.null(x$nsim))
     cat("\n- based on ", x$nsim,
-        " simulation", if (x$nsim > 1) "s", "\n",
+        " repetition", if (x$nsim > 1) "s", "\n",
         sep = "")